Add files via upload
Fourier analysis lays the foundations for many methods used in the field of digital signal processing))) The Fourier transform (in fact, there are several variants of such transformations) allows you to match a signal given in the time domain with its equivalent representation in the frequency domain. On the contrary, if the frequency response of the signal is known, then the inverse Fourier transform allows you to determine the corresponding signal in the time domain.